wait-signal
This (unimplemented) macro action will block until the (also
unimplemented) send-signal macro action is executed by some other macro.
This would not be possible if we had to completely expand each macro as
we encounted them, but it is possible in our implementation because
we split the work into a number of smaller steps, which can be
interleaved with the expansion steps of some other macros. This way,
once a macro gets stuck waiting for a signal, we can switch our focus to
the steps of a different macro, which may send the required signal.

Exchanging signals does not currently exchange any
computationally-useful information, and so they cannot be used to
implement any useful macros. They only serve as a proof of concept
illustrating that it is indeed possible for a macro to get stuck, and to
resume after some other condition gets fullfilled.