simplify Set manipulation using Control.Lens.non
With nested Maps, a common action is to insert an item at k1 in the
outer map and k2 in the inner Map. If the outer map does not have an
entry for k1 yet, we typically have to write some extra logic to
initialize that entry to the empty Map before we can insert k2 into it.

Lenses are supposed to make it easy to manipulate nested data structure,
but at first glance it is not clear how to index into nested Maps:

at k1 :: Lens' (Map k1 (Map k2 a))
               (Maybe (Map k2 a))

Due to the Maybe, we cannot simply chain at's, we must somehow strip off
the Maybe first. The little-known lens called "non" does that job:

non Map.empty :: Lens' (Maybe (Map k2 a))
                       (Map k2 a)
at k1 . non Map.empty :: Lens' (Map k1 (Map k2 a))
                               (Map k2 a)

The way it works is that it interprets Nothing (that is, k1 is not in
the outer map) as a Map.empty; if we use "over" or "set" to turn that
into a non-empty Map, "non" will wrap that non-empty Map in a Just,
causing the "at k1" to insert the map at the key k1. This way, we don't
have to write some extra logic to initialize the k1 entry to the empty
Map before we can insert k2 into it.

"non" also cleans up the outer Map after deletions. If the k1 entry
already contains a Map, and we use "over" or "set" to delete an entry in
that Map, and it just so happened that it was the last entry of that
inner Map which we deleted, then the inner Map is now "Map.empty".
"non Map.empty" detects this and automatically removes the k1 entry from
the outer Map, because a missing entry means the same thing as an entry
containing an empty Map.