begin a commentary
The codebase is already so complex that I have a hard time contributing
to it, and I was on the project since the very beginnning! I think it
would be helpful to have a "commentary", that is, some long form
document explaining the overall architecture of the code. The name is
obviously inspired by the GHC Commentary, but I intend to make it a
separate markdown document rather than a bunch of separate comments
distributed throughout the codebase.