distinguish between Show and ShortShow
Show's output should include all the fields, formatted in a way that we
could paste that output in ghci to reconstruct the value.

ShortShow's output may omit fields and constructors in order to improve
readability. The format should look like valid Haskell code (that is,
the output is still intended for programmers and auto-formatters, not
users), but pasting it in ghci is likely to fail because of the missing
fields.