separate already-added vs to-be-added ScopeSets
Previously, we were threading a ScopeSet through all the declarations in
a declaration group, and at each step, we were adding all of the scopes
so far to the remaining syntax. This means that the same Scope was
redundantly added to the same syntax many times.

This new design distinguishes the scopes which have already been added
to the syntax in the rest of a declaration group from the scopes which
have not yet been added. As a result, most declarations no longer need
to thread that ScopeSet through, as they only need to record the scopes
they introduce themselves, not those which have been introduced by
earlier declarations. Only the ExpandDeclForms task needs to worry about
threading the scopes through.

The previous version also used the word "environment" incorrectly in the
comments explaining how these ScopeSets are threaded via the
DeclValidityPtrs. The new version of those comments talk about scopes,
not environments.