ppBind only returns the newly-bound vars
The previous version of this code incorrectly assumed that bbBind
returned and _extended_ environment in which the newly-bound vars are
added to the input environement. As a result, it threaded the
environment through each decl in the list.

But in fact, bbBind returns an environment which only contains the
newly-bound vars, and so things are a bit more complex. We still thread
an environment through the computation in order to accumulate all the
variables bound by each decl in the list, but this environement must not
start with the input environment, it must start empty. Each decl can see
the input environment _and_ the variables bound by the previous decls,
so we must combine those two environments in the recursive call.