Utility macros for matching on particular identifiers
Two new utility macros:

mcond is a monadic version of Lisp's cond operator, in which the tests
are expected to have type (Macro Bool).

free-identifier-case is a case operator on identifiers, useful for
testing against a few different special-cased operators without having
to nest a bunch of ifs and binds and free-identifier=?s.