Working type-case and stuck macros
This commit contains the final bits of run-time and expander support
for a monadic type-case operator that gets stuck on unsolved metas.

The which-problem example has been updated with a demonstration that
the type checker's order of traversal is independent of the
availability of type information to the macro expander.