every type has an argument list
possibly an empty one. in practice, the type constructor determines the
length of that list, so TFun always has 2 arguments and TString always
has zero. thus, this more-uniform representation is less precise than
before, in that it makes more illegal states representable, but it paves
the way for the implementation of higher-kinded types. for example, it
is now possible to represent TFun applied to zero or one argument;
currently, writing either of those in a Klister program is a kind error,
but soon we can make those types of kind * -> * -> * and * -> *.