better golden test output upon failure
Previously we were displaying the incorrect output using the default
format, which is very hard to read:

  "#<closure> : \226\136\128(\206\177 : *). (Syntax \226\134\146 \206\177)\n"

Clearly, those escaped characters are supposed to be unicode characters.
Unfortunately, loading that string in ghci and printing it out does not
give the expected result:

  λ> s = "#<closure> : \226\136\128(\206\177 : *). (Syntax \226\134\146 \206\177)\n"
  λ> putStrLn s
  #<closure> : â<88><80>(Î± : *). (Syntax â<86><92> Î±)

That's because this isn't a String, but a ByteString! We need to decode
it into UTF8 in order to get something readable:

  λ> b = ("#<closure> : \226\136\128(\206\177 : *). (Syntax \226\134\146 \206\177)\n" :: ByteString)
  λ> Text.putStrLn $ decodeUtf8 b
  #<closure> : ∀(α : *). (Syntax → α)

With this commit, we now display the incorrect output using much more
readable unicode, and instead of displaying the entire output, we now
display it as a diff between the expected and actual output, e.g.

  -#<closure> : ∀α. (Syntax → α)
  +#<closure> : ∀(α : *). (Syntax → α)