polish ambiguous-identifier error message
whenever we encounter this error message, we typically temporarily
extend the error message to include the scope sets in order to make
debugging easier. that's a sign that this information should probably be
permanently-included in the error message.

I added some text explaining what each part of the error message is
about, I have tweaked the layout so that the scope sets all start on the
same column, and I have changed the format of the identifier's scope
sets to match that of the candidates; that format is less pretty, but
it's easier to diff it with the other scope set lines.