Add deep pattern matching to core language
Fixes #107.

Macros in pattern position complicate the semantics a little bit,
because pattern variables are lexically identical to any other kind of
identifier. There's no easy way to tell whether a user intends to
shadow an existing macro with a pattern variable or to invoke that
macro in a pattern position.

The solution here is perhaps overly simplistic: bare identifiers in
pattern position are ALWAYS pattern variables - even if they share a
name with a constructor or a macro. The rule is simple, but not very
flexible.

Other alternative approaches could include:

 1. Lexically distinguish pattern variables from other identifiers,
 e.g. by wrapping them in <> or requiring them to start or end with a
 given character. This is the Haskell/Idris approach.

 2. Have some more rules to check for exisiting bindings, and specify
 when they override the local binding. This is the Agda approach.

 3. Allow macros to specify at binding time whether they should be
 considered in pattern contexts, or not. This is the Racket approach (see
 define-match-expander).

I think that any of those can build on top of this basic run-time
support for deep patterns, as well as the expander's handling of
bindings and the like. Our experience in Idris, which went from option
2 to option 1 a few years back, was that simple rules were better than
more flexible yet more complicated rules.