no more , magic
previously, intermediate-syntax-case was generating a `(...) expression,
and thus right-hand-sides could use ,(...) expressions to splice some
code. that was strange because those ,(...) were not lexically
surrounded by `(...). Now each right-hand-side is a Racket expression,
not a Klister expression, and must thus use `(...) itself it it wants to
return a Klister expression.