don't use the (,@(list xs ...) x) pattern
given a list of N+1 elements, it assigns the first N to xs and the last
one to x. This is convenient, but it doesn't generalize well: what if we
had (,@(list xs ...) ,@(list ys ...)), which fraction should be assigned
to which side? for this reason, generate-syntax-parse doesn't support
this kind of pattern, and thus we cannot use it to implement
fancy-syntax-parse. since generate-syntax-parse is used as a template
demonstrating how to implement such a function, I've updated
generated-syntax-parse to use the more reasonable pattern (x ,@xs).