auto-splice
simplifies code-generation. another alternative would have been to
define, for each Klister form, a Racket macro which mimics that form and
outputs racket code which constructs the Klister code for that form by
taking Racket expressions producing the code for the arguments and
splicing the results in the generated code for the form. That would
require work proportional to the number of syntax forms in Klister,
whereas the auto-splice approach only requires work proportional to the
number of generate- functions defined in bootstrap.rkt, which is
currently much smaller.