display expected/got on multiple lines
I initially used the Show encoding in order to debug a problem with
goldenVsString, but just reimplementing goldenVsString in this way has
managed to fix the problem, so I no longer need this more detailed but
harder-to-read encoding. Let's revert to the original encoding, while
still using my reimplementation of goldenVsString.

Also, there is no reason to convert everything to ByteStrings anymore, I
was only doing that in order to match goldenVsString's API.