hush 'incomplete' pattern-matches
they are incomplete in the sense that they don't match on all Core
values, but complete in the sense that they cover all the constructors
of the expected type. so if those impossible lines are ever reached, it
means there is a bug in the type checker allowing Core values of the
wrong type to reach these primitives.