bugfix: bind-IO only ran the first IO action!
the issue was that given (bind-IO iox f), we were basically executing

    do x <- iox
       pure (f x)

instead of

    do x <- iox
       f x

It's actually worse than that: we were returning the IO action returned
by (f x) to a caller which expects a y, not an (IO y). So this could
easily have crashed the program, if we ever wrote a program which
actually did something with the result of bind-IO.

In practice, however, most of our tests are about expressions, not IO
actions, so we didn't catch this earlier. The one test which did use
bind-IO was io.kl, but it only checked that the result was inferred to
have type IO Integer, it did not execute the resulting IO action.