README rewrite
The previous version didn't even explain that Klister is a programming
language, this was very confusing for visitors unfamiliar with the
project.

In the new version, I state right away that Klister is a programming
language, and provide a short example program highlighting some of the
language features. I foreshadows what makes Klister interesting, to keep
the reader interested, without frontloading the document with all the
background knowledge which would be necessary to understand stuck
macros.

I then list the features of the language, starting with some
easier-to-understand features which readers are familiar with from other
languages. I would like each feature to link to some commented code
example demonstrating the feature, but many of those links are still
TODO.

Then, I finally list the features which make Klister special, but are a
bit harder to understand because almost no other language has them.

Finally, to counter-balance all that positivity from a long list of
features which Klister _does_ have, I help the reader decide if Klister
is for them by admitting to some important features which Klister does
_not_ have.

I moved the implementation details to a file called "HACKING.rst", as I
believe this is the traditional name for a file explaining how the
codebase is organized.