round braces
The code with square braces was written in an earlier version of Klister
in which square brackets and round parentheses had different semantics.
Today, they have the same semantics. I thus rewrote the parentheses to
follow the convention that round parentheses are used by default, and
square brackets are only used to wrap a left-hand-side and a
right-hand-side, to make it clearer that the two positions have
completely different meanings.