combine baseType, unaryType, arrowType, addDatatype...
...and addDatatypePrimitive.

The main way in which they differ is the number of type arguments they
accept, and the kind of those arguments. A single function parameterized
by a list of kinds can thus be specialized to those various special
cases.

Another, less important difference, is that unaryType does not require
parentheses around the type name, e.g. Syntax, not (Syntax). A
consequence of this unification is that primitive types such as Syntax
are now written (Syntax), just like datatypes with no type parameters.