add use-site scope to every use site
As per https://users.cs.utah.edu/plt/scope-sets/pattern-macros.html#%28part._use-site%29

We were adding a scope to every identifier introduced by the macro
definition, but we were not adding a scope to every identifier
introduced by the call site; as a result, it was possible for an
identifier introduced by the macro to refer to an identifier introduced
by the call-site. The two sets of identifiers are now completely
independent.