simplify lambda-case example
the previous version was using define-macros, which is a bit distracting
because we are only defining a single macro. The "define-syntax-rule.kl"
module provides the less-distractingly-named "define-macro", but it is
designed for macros which take a fixed number of arguments. Since
lambda-case takes a variable number of cases, we have to use the
still-distractingly-named "define-variadic-macro". But at least the code
is now shorter.