Pattern, TypePattern, vs PolyProblem prims
previously, "else" was considered a "pattern primitive", even though it
is valid in both the pattern and the type pattern Problems. And
"with-unknown-macro" was considered a "universal primitive", because it
works in any Problem.

Now, a "pattern primitive" means a pattern which is _only_ valid in the
pattern Problem, and a "type pattern primitive" (a new term!) means a
pattern which is _only_ valid in the type pattern Problem. There are no
such primitives yet, but I plan to add a type pattern primitive, this
commit is in preparation for that.

Finally, a "poly-problem primitive" now means a primitive which works in
more than one Problem; all of them in the case of "with-unknown-type",
and two of them in the case of "else".