fork job at the phase of the output
forkAwaitingTypeCase schedules a job which (eventually) executes the
Macro actions which follow the type-case. shifting the phase before
forking did cause that Macro action to be executed in a shifted phase,
but also had the unintended effect of shifting the phase at which to
interpret the Syntax returned by the macro. The solution: don't shift
the fork call, but do shift the Macro action it eventually executes.