new Problem: type constructor
I want to add a type pattern macro in phase 1 which matches on a type
constructor in phase 0. The only way to lookup a type constructor is to
expand the name of that type constructor (so that hygiene rules apply
properly). That means we now need macros which return type constructors,
not types nor type patterns.