hide caller env from callee
in a lambda, only the variables which are lexically visible at the
definition site are visible, not the variables which are lexically
visible at the call site. therefore, when invoking a closure, we should
evaluate the body in the captured environment, not in a combined
environment containing both the captured environment and the call site's
environment.