always use the captured environment
now that the caller's variables are no longer included in the
environment, this exposed a bug where the evaluator recurred with the
upward-propagated environment rather than the environment which was
captured in the Kont. actually, it never makes sense to be propagating
the callee's environment to the caller, so all the other uses of that
environment must also be bugs. let's not propagate this environment at
all.